Rename folder: providers -> state-providers
