+++ frontend/src/managers/panic.ts
@@ -1,7 +1,7 @@
 import { TextButtonWidget } from "@/components/widgets/buttons/TextButton";
 import { Editor } from "@/interop/editor";
 import { DisplayDialogPanic, WidgetLayout } from "@/interop/messages";
-import { DialogState } from "@/providers/dialog";
+import { DialogState } from "@/state-providers/dialog";
 import { stripIndents } from "@/utilities/strip-indents";
 
 export function createPanicManager(editor: Editor, dialogState: DialogState): void {
